requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
sqlalchemy==2.0.28
pytest==8.0.2
httpx==0.27.0